lots of progress on fully functional config stuff

+++ src/java/com/rapleaf/hank/coordinator/HostDomainConfig.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.util.Set;
 
 public interface HostDomainConfig {
-  public byte getDomainId();
+  public int getDomainId();
 
   public Set<HostDomainPartitionConfig> getPartitions() throws IOException;
